+++ tests/handlers/articles/alis_tokens/show/test_articles_alis_tokens_show.py
@@ -1,5 +1,5 @@
 from unittest import TestCase
-from article_content_show import ArticleContentShow
+from articles_alis_tokens_show import ArticlesAlisTokensShow
 from unittest.mock import patch, MagicMock
 import yaml
 import os
@@ -7,15 +7,15 @@ import boto3
 import json
 
 
-class TestArticleContnetShow(TestCase):
+class TestArticlesAlisTokensShow(TestCase):
     dynamodb = boto3.resource('dynamodb', endpoint_url='http://localhost:4569/')
 
-    target_tables = ['ArticleInfo', 'ArticleContent']
+    target_tables = ['ArticleAlisToken', 'ArticleEvaluatedManage']
 
     @classmethod
     def setUpClass(cls):
-        os.environ['ARTICLE_INFO_TABLE_NAME'] = 'ArticleInfo'
-        os.environ['ARTICLE_CONTENT_TABLE_NAME'] = 'ArticleContent'
+        os.environ['ARTICLE_ALIS_TOKEN_TABLE_NAME'] = 'ArticleAlisToken'
+        os.environ['ARTICLE_EVALUATED_MANAGE_TABLE_NAME'] = 'ArticleEvaluatedManage'
 
         f = open("./database.yaml", "r+")
         template = yaml.load(f)
@@ -27,41 +27,35 @@ class TestArticleContnetShow(TestCase):
 
             cls.dynamodb.create_table(**create_params)
 
-        article_info_table = cls.dynamodb.Table('ArticleInfo')
-        items = [
+        article_evaluated_manage_table = cls.dynamodb.Table('ArticleEvaluatedManage')
+        article_evaluated_manage_table.put_item(Item={'active_evaluated_at': 1520150272000000})
+
+        article_alis_token_table = cls.dynamodb.Table('ArticleAlisToken')
+        article_alis_tokens = [
             {
                 'article_id': 'testid000001',
-                'status': 'published',
-                'title': 'testid000001 titile',
-                'sort_key': 1520150272000000
+                'alis_token': 100,
+                'evaluated_at': 1520150272000000
             },
             {
                 'article_id': 'testid000002',
-                'status': 'published',
-                'title': 'testid000002 titile',
-                'sort_key': 1520150272000001
-            }
-        ]
-
-        for item in items:
-            article_info_table.put_item(Item=item)
-
-        article_content_table = cls.dynamodb.Table('ArticleContent')
-        items = [
-            {
-                'article_id': 'testid000001',
-                'body': 'testid000001 body',
-                'title': 'testid000001 titile'
+                'alis_token': 50,
+                'evaluated_at': 1520150272000000
             },
             {
                 'article_id': 'testid000003',
-                'body': 'testid000003 body',
-                'title': 'testid000003 titile'
+                'alis_token': 150,
+                'evaluated_at': 1520150572000000
+            },
+            {
+                'article_id': 'testid000001',
+                'alis_token': 80,
+                'evaluated_at': 1520150572000000
             }
         ]
 
-        for item in items:
-            article_content_table.put_item(Item=item)
+        for article_alis_token in article_alis_tokens:
+            article_alis_token_table.put_item(Item=article_alis_token)
 
     @classmethod
     def tearDownClass(cls):
@@ -69,7 +63,7 @@ class TestArticleContnetShow(TestCase):
             cls.dynamodb.Table(table_name).delete()
 
     def assert_bad_request(self, params):
-        function = ArticleContentShow(params, {}, self.dynamodb)
+        function = ArticlesAlisTokensShow(params, {}, self.dynamodb)
         response = function.main()
 
         self.assertEqual(response['statusCode'], 400)
@@ -81,20 +75,18 @@ class TestArticleContnetShow(TestCase):
             }
         }
 
-        response = ArticleContentShow(params, {}, self.dynamodb).main()
+        response = ArticlesAlisTokensShow(params, {}, self.dynamodb).main()
 
         expected_item = {
             'article_id': 'testid000001',
-            'status': 'published',
-            'title': 'testid000001 titile',
-            'body': 'testid000001 body',
-            'sort_key': 1520150272000000
+            'alis_token': 100,
+            'evaluated_at': 1520150272000000
         }
 
         self.assertEqual(response['statusCode'], 200)
         self.assertEqual(json.loads(response['body']), expected_item)
 
-    @patch('article_content_show.validate', MagicMock(side_effect=Exception()))
+    @patch('articles_alis_tokens_show.validate', MagicMock(side_effect=Exception()))
     def test_main_ng_with_internal_server_error(self):
         params = {
             'pathParameters': {
@@ -102,29 +94,18 @@ class TestArticleContnetShow(TestCase):
             }
         }
 
-        response = ArticleContentShow(params, {}, self.dynamodb).main()
+        response = ArticlesAlisTokensShow(params, {}, self.dynamodb).main()
 
         self.assertEqual(response['statusCode'], 500)
 
-    def test_article_info_record_not_found(self):
+    def test_record_not_found(self):
         params = {
             'pathParameters': {
                 'article_id': 'testid000003'
             }
         }
 
-        response = ArticleContentShow(params, {}, self.dynamodb).main()
-
-        self.assertEqual(response['statusCode'], 404)
-
-    def test_article_content_record_not_found(self):
-        params = {
-            'pathParameters': {
-                'article_id': 'testid000002'
-            }
-        }
-
-        response = ArticleContentShow(params, {}, self.dynamodb).main()
+        response = ArticlesAlisTokensShow(params, {}, self.dynamodb).main()
 
         self.assertEqual(response['statusCode'], 404)
